@@ -2,21 +2,32 @@
   "name": "poka",
   "version": "1.0.0",
   "description": "",
-  "main": "bot.js",
+  "main": "./dist/bot.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node bot.js"
+    "build": "tsc",
+    "dev": "nodemon ./src/bot.ts"
   },
   "author": "Freedzio",
   "license": "ISC",
   "dependencies": {
+    "@discordjs/opus": "^0.6.0",
     "discord.js": "^12.5.1",
     "dotenv": "^8.2.0",
-    "node-fetch": "^2.6.1",
-    "nodemon": "^2.0.7"
+    "ffmpeg": "0.0.4",
+    "ffmpeg-static": "^4.4.0",
+    "fluent-ffmpeg": "^2.1.2",
+    "node-fetch": "2.6.5",
+    "ts-node-dev": "^1.1.8",
+    "ytdl-core": "^4.9.1",
+    "ytsr": "^3.5.3"
   },
   "engines": {
-    "node": "12.18.1",
+    "node": ">=14.17.3",
     "npm": "*"
+  },
+  "devDependencies": {
+    "@types/node": "^16.11.0",
+    "@types/node-fetch": "^3.0.3"
   }
 }
